36 fix alembic script migration

+++ alembic/versions/a90939222d2a_create_recipes_table.py
@@ -21,8 +21,8 @@ def upgrade() -> None:
         'recipes',
         sa.Column('recipe_id', sa.Integer, primary_key=True),
         sa.Column('recipe_name', sa.String(50), nullable=False),
-        sa.Column('calories', sa.Integer(50), nullable=True),
-        sa.Column('prep_time_mins', sa.Integer(50), nullable=True),
+        sa.Column('calories', sa.Integer, nullable=True),
+        sa.Column('prep_time_mins', sa.Integer, nullable=True),
         sa.Column('recipe_instructions', sa.String(50), nullable=True),
         sa.Column('recipe_url', sa.String(50), nullable=True),
     )
